+++ test_autofit/mapper/test_by_path.py
@@ -1,7 +1,7 @@
 import pytest
 
 import autofit as af
-from autofit.mock.mock_model import MockWithTuple
+from autofit.mapper.mock.mock_model import MockWithTuple
 from autofit.non_linear.mock.mock_search import MockSearch
 
 @pytest.fixture(
