+++ test_autofit/mapper/test_has.py
@@ -1,6 +1,6 @@
 import autofit as af
 
-from autofit.mock.mock_model import MockClassx2
+from autofit.mapper.mock.mock_model import MockClassx2
 
 class GaussianChild(af.Gaussian):
     pass
