+++ test_autofit/mapper/with_paths/test_samples.py
@@ -1,7 +1,7 @@
 import pytest
 
 import autofit as af
-from autofit.mock.mock_model import MockWithTuple
+from autofit.mapper.mock.mock_model import MockWithTuple
 
 
 @pytest.fixture(
