+++ test_autofit/non_linear/nest/test_dynesty.py
@@ -5,7 +5,7 @@ import pytest
 
 import autofit as af
 
-from autofit.mock.mock_model import MockClassx4
+from autofit.mapper.mock.mock_model import MockClassx4
 
 pytestmark = pytest.mark.filterwarnings("ignore::FutureWarning")
