+++ test_autofit/non_linear/samples/test_nest.py
@@ -2,7 +2,7 @@ import pytest
 
 import autofit as af
 
-from autofit.mock.mock_model import MockClassx4
+from autofit.mapper.mock.mock_model import MockClassx4
 from autofit.non_linear.mock.mock_samples import MockNestSamples
 
 pytestmark = pytest.mark.filterwarnings("ignore::FutureWarning")
