+++ test_autofit/non_linear/samples/test_pdf.py
@@ -4,7 +4,7 @@ import pytest
 from autoconf.conf import with_config
 
 import autofit as af
-from autofit.mock.mock_model import MockClassx2, MockClassx4
+from autofit.mapper.mock.mock_model import MockClassx2, MockClassx4
 from autofit.non_linear.mock.mock_samples import MockSamples
 
 pytestmark = pytest.mark.filterwarnings("ignore::FutureWarning")
