+++ test_autofit/non_linear/samples/test_samples.py
@@ -2,7 +2,7 @@ import os
 import pytest
 
 import autofit as af
-from autofit.mock.mock_model import MockClassx2, MockClassx4
+from autofit.mapper.mock.mock_model import MockClassx2, MockClassx4
 from autofit.non_linear.mock.mock_samples import MockSamples
 
 pytestmark = pytest.mark.filterwarnings("ignore::FutureWarning")
