+++ test_autofit/non_linear/test_abstract_search.py
@@ -8,7 +8,7 @@ import pytest
 import autofit as af
 from autoconf import conf
 
-from autofit.mock.mock_model import MockClassx4, MockClassx2Tuple
+from autofit.mapper.mock.mock_model import MockClassx4, MockClassx2Tuple
 from autofit.non_linear.mock.mock_search import MockSearch, MockSamples
 
 pytestmark = pytest.mark.filterwarnings("ignore::FutureWarning")
