+++ test_autofit/non_linear/test_initializer.py
@@ -1,5 +1,5 @@
 import autofit as af
-from autofit.mock.mock_model import MockClassx4
+from autofit.mapper.mock.mock_model import MockClassx4
 
 
 class MockFitness:
