+++ test_autofit/non_linear/test_result.py
@@ -1,7 +1,7 @@
 import pytest
 import autofit as af
 
-from autofit.mock.mock_model import MockClassx2Tuple
+from autofit.mapper.mock.mock_model import MockClassx2Tuple
 from autofit.non_linear.mock.mock_samples import MockSamples
 from autofit.non_linear.mock.mock_search import MockSearch
